Change MCMC and HMC loop counters to int32 to reduce CPU-GPU communication in TF graph mode.
Also eliminate redundant tf.convert_to_tensor from smart_for_loop,
thus generalizing the dtype the upstream user can use to spell the
number of iterations.

Together with cl/241853748, this speeds up an informal HMC benchmark
from 16.6s to 11.7s (about 30%) on GPU in TF graph mode.  The
benchmark program is a Bayesian linear regression, 30 latent
dimensions, 10 chains in batch, 1000 iterations of 30-leapfrog-step
HMC, on 30,000 synthetic data points, tracing everything, no burn-in,
no thinning.

PiperOrigin-RevId: 242534828

+++ experimental/neutra/neutra_kernel.py
@@ -311,7 +311,7 @@ class NeuTra(tfp.mcmc.TransitionKernel):
   def _num_leapfrog_steps(self, step_size):
     step_size = tf.convert_to_tensor(value=step_size)
     trajectory_length = np.float32(self._total_event_size)**0.25
-    return tf.cast(tf.math.ceil(trajectory_length / step_size), dtype=tf.int64)
+    return tf.cast(tf.math.ceil(trajectory_length / step_size), dtype=tf.int32)
 
   def _flatten_state(self, state):
     state_parts = tf.nest.flatten(state)
